Refactor Search Form in Screncasts with SearchComponent
Description:
This commit replaces the search form in app/views/screencasts/index.html.erb with the SearchComponent. The SearchComponent encapsulates the search form logic, making the code more readable and easier to change. It also takes care of HTML escaping the search_term, improving security. The form is now rendered inside a <div> instead of a <span> to follow HTML best practices.